Fix Pages Build Job
The CI build job for pages was never updated during the documentation redo, and it was never tested, so it was failing. This PR fixes that and makes sure the pages build always runs so we knows if it is failing. The pages deploy only runs on tagged builds. This also allows the docs to build and deploy in parallel.

MR !14 squashed commit

Signed-off-by: Max Ehrlich <max.ehr@gmail.com>